refactor(chalice): moved CH sessions-search to FOSS refactor(DB): changed DB structures for CH sessions-search in FOSS refactor(DB): preparing for v1.23.0
