+++ api/env.default
@@ -75,3 +75,4 @@ EXP_AUTOCOMPLETE=true
 EXP_ALERTS=true
 EXP_ERRORS_SEARCH=true
 EXP_METRICS=true
+EXP_SESSIONS_SEARCH=true
