+++ ee/scripts/schema/db/init_dbs/clickhouse/1.23.0/1.23.0.sql
@@ -1,0 +1,1 @@
+CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.23.0-ee';
