+++ ee/scripts/schema/db/init_dbs/clickhouse/create/init_schema.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.22.0-ee';
+CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.23.0-ee';
 CREATE DATABASE IF NOT EXISTS experimental;
 
 CREATE TABLE IF NOT EXISTS experimental.autocomplete
