+++ ee/scripts/schema/db/rollback_dbs/clickhouse/1.23.0/1.23.0.sql
@@ -1,0 +1,1 @@
+CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.22.0-ee';
