+++ scripts/schema/db/init_dbs/clickhouse/create/init_schema.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.22.0';
+CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.23.0';
 CREATE DATABASE IF NOT EXISTS experimental;
 
 CREATE TABLE IF NOT EXISTS experimental.autocomplete
@@ -532,3 +532,15 @@ CREATE TABLE IF NOT EXISTS product_analytics.all_properties
     _timestamp        DateTime DEFAULT now()
 ) ENGINE = ReplacingMergeTree(_timestamp)
       ORDER BY (project_id, property_name, is_event_property);
+
+
+CREATE TABLE IF NOT EXISTS experimental.user_viewed_sessions
+(
+    project_id UInt16,
+    user_id    UInt32,
+    session_id UInt64,
+    _timestamp DateTime DEFAULT now()
+) ENGINE = ReplacingMergeTree(_timestamp)
+      PARTITION BY toYYYYMM(_timestamp)
+      ORDER BY (project_id, user_id, session_id)
+      TTL _timestamp + INTERVAL 3 MONTH;
