+++ scripts/schema/db/rollback_dbs/clickhouse/1.23.0/1.23.0.sql
@@ -1,0 +1,4 @@
+CREATE OR REPLACE FUNCTION openreplay_version AS() -> 'v1.22.0';
+
+
+DROP TABLE IF EXISTS experimental.user_viewed_sessions;
